Update questions after AI re-run instead of replacing
Now that we can attach answers to questions, it makes sense to still be
able to re-run the AI (eg. with a newer model) but we don't want to lose
the data associated with the question.

We now just update the question and answer columns with the new
extraction result instead of deleting and replacing.

We've also renamed the persistence layer function from
`markQuizAIExtractionCompleted` to
`upsertQuizQuestionsWithSuccessfulAIExtraction` to better describe what
it actually does.